build(docz): freeze version at 0.12.5
There's a bug in some later versions of docz where the menu scrim covers the menu on narrow screens,
and the even later versions where it's fixed also doesn't work correctly, because of some bug that
makes the whole docz-app crash. So freeze the version at 0.12.5, and only upgrade when a new version
is confirmed to work with narrow screens.