refactor(platform): add get method in favor of public properties
The old interface was hard to maintain since it violated both the open-closed principle and the
interface segregation principle. Since it for all intents and purposes is a service locator, it
makes more sense to have a clean interface that allows us to add new services without introducing a
breaking change.

BREAKING CHANGE: Removed all public properties, services are now retrieved with the `get` method.
New services can be registered with the `register` method which can also be used by 3rd parties to
register new services on the platform. Core platform service names are exposed in the
`PlatformServiceName` enum.