chore(package): update dependency on typedoc-plugin-markdown to include patch-versions
The developer of this plugin releases new patch versions every
other day, and they don't seem to actually fix any bugs. It seems
like he's not following semantic versioning, and he's not keeping
a changelog. With this change, Greenkeeper will only open a PR
for new minor-versions, which should be plenty for us.